10503 - moved toggle component higher in the hierarchy to SearchSidebar.jsx
